Move callgraph extension back into semanticcpg.
