+++ cpgserver/src/main/scala/io/shiftleft/cpgserver/CpgServerController.scala
@@ -100,7 +100,7 @@ class CpgServerController(impl: ServerImpl, system: ActorSystem = ActorSystem())
       e.eval(s"""
                 import io.shiftleft.codepropertygraph.Cpg
                 | import io.shiftleft.semanticcpg.language._
-                | import io.shiftleft.callgraph.language.NoResolve
+                | import io.shiftleft.semanticcpg.language.NoResolve
                 | implicit val resolver = NoResolve
                 | val cpg = aCpg.asInstanceOf[io.shiftleft.codepropertygraph.Cpg]
                 | $query
