+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/callgraphextension/Call.scala
@@ -1,6 +1,5 @@
-package io.shiftleft.callgraph.language.extensions
+package io.shiftleft.semanticcpg.language.callgraphextension
 
-import io.shiftleft.callgraph.language.ICallResolver
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, nodes}
 import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.language.types.expressions.{Call => OriginalCall}
