+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/callgraphextension/Method.scala
@@ -1,7 +1,6 @@
-package io.shiftleft.callgraph.language.extensions
+package io.shiftleft.semanticcpg.language.callgraphextension
 
 import gremlin.scala._
-import io.shiftleft.callgraph.language.ICallResolver
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, nodes}
 import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.language.types.expressions.{Call => OriginalCall}
