+++ semanticcpg/src/test/scala/io/shiftleft/semanticcpg/language/CallGraphTests.scala
@@ -1,6 +1,5 @@
-package io.shiftleft.callgraph.language
+package io.shiftleft.semanticcpg.language
 
-import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.testfixtures.CodeToCpgFixture
 import org.scalatest.{Matchers, WordSpec}
